pyproject.toml
[tool.poetry]
name = "coolify-demo"
version = "0.1.0"
description = ""
authors = ["Iliyan Vutoff <iliyan@dext.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
db-psycopg2 = "^0.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
